requirements.txt
streamlit
beautifulsoup4
requests
selenium
webdriver-manager
chromadb
tiktoken
python-dotenv
duckdb
pypdf
google-genai
